compile warning fix.

+++ src/server/game/Battlegrounds/SpectatorAddon.cpp
@@ -101,7 +101,7 @@ WorldPacket& SpectatorAddonMsg::CachedMessage()
                 {
                     case SPECTATOR_PREFIX_PLAYER:
                         char buffer[64];
-                        sprintf(buffer, "%i;", player);
+                        sprintf(buffer, "%i;", (int)player);
                         msg += buffer;
                         break;
                     case SPECTATOR_PREFIX_NAME:
@@ -110,7 +110,7 @@ WorldPacket& SpectatorAddonMsg::CachedMessage()
                     case SPECTATOR_PREFIX_TARGET:
                     {
                         char buffer[20];
-                        sprintf(buffer, "TRG=%i;", target);
+                        sprintf(buffer, "TRG=%i;", (int)target);
                         msg += buffer;
                         break;
                     }
